[SCP-3233] Added interface to cloudflare.json. This configures the additional fields we need for domains.
The user can specify a name and phone number which is sent with domain registration. Since we associated this internally to an account, if multiple domains are added these values will replace any existing values.

There is a hidden client ID field which will be returned by our system after a domain is added. This will be stored in the configuration on the CloudFlare side and will be used to inject the correct URL for loading the Sociocast javascript file.